requirements.txt
selenium
pytest
pytest-xdist
pytest-rerunfailures
python-dotenv